// src/api/models/PaisDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PaisDto = {
    id?: number;
    nombre?: string | null;
    codigoISO?: string | null;
    codigoTelefonico?: string | null;
    activo?: boolean;
    fechaCreacion?: string;
    fechaModificacion?: string | null;
};